Extract PersonalNoteExport.ExportTask
